chore(workspace): :recycle: moved 'OmitId' type into 'UtilityTypes' file
